+++ passes/S018/testdata/src/a/comment_ignore.go
@@ -1,0 +1,21 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform/helper/schema"
+)
+
+func fcommentignore() {
+	//lintignore:S018
+		_ = schema.Schema{
+		MaxItems: 1,
+		Type:     schema.TypeSet,
+	}
+
+	//lintignore:S018
+	_ = map[string]*schema.Schema{
+		"name": {
+			MaxItems: 1,
+			Type:     schema.TypeSet,
+		},
+	}
+}